+++ elastic-job-core/src/test/java/com/dangdang/ddframe/job/internal/election/ElectionListenerManagerTest.java
@@ -95,15 +95,4 @@ public final class ElectionListenerManagerTest {
         verify(leaderElectionService).leaderElection();
         verify(shardingService).setReshardingFlag();
     }
-    
-//    @Test
-//    public void assertListenLeaderElection() {
-//        getRegistryCenter().persist("/testJob/leader/election/host", "host0");
-//        assertFalse(getRegistryCenter().isExisted("/testJob/leader/sharding/necessary"));
-//        WaitingUtils.waitingShortTime();
-//        getRegistryCenter().remove("/testJob/leader/election");
-//        WaitingUtils.waitingShortTime();
-//        assertThat(getRegistryCenter().get("/testJob/leader/election/host"), is(localHostService.getIp()));
-//        assertTrue(getRegistryCenter().isExisted("/testJob/leader/sharding/necessary"));
-//    }
 }